refactor xwalk
