+++ portality/api/v1/crud/applications.py
@@ -2,13 +2,13 @@ from portality.api.v1.crud.common import CrudApi
 from portality.api.v1 import Api401Error, Api400Error, Api404Error, Api403Error, Api409Error
 from portality.api.v1.data_objects import IncomingApplication, OutgoingApplication
 from portality.lib import dataobj
-from datetime import datetime
 from portality import models
 
 from portality.bll import DOAJ
 from portality.bll.exceptions import AuthoriseException, NoSuchObjectException
 from portality import lock
-from portality.formcontext import formcontext, xwalk
+from portality.formcontext import formcontext
+from portality.formcontext.xwalks import suggestion_form
 from werkzeug.datastructures import MultiDict
 
 from copy import deepcopy
@@ -96,7 +96,7 @@ class ApplicationsCrudApi(CrudApi):
                 raise Api404Error()
 
             # convert the incoming application into the web form
-            form = MultiDict(xwalk.SuggestionFormXWalk.obj2form(ap))
+            form = MultiDict(suggestion_form.SuggestionFormXWalk.obj2form(ap))
 
             fc = formcontext.ApplicationFormFactory.get_form_context(role="publisher", form_data=form, source=vanilla_ap)
             if fc.validate():
@@ -117,7 +117,7 @@ class ApplicationsCrudApi(CrudApi):
         # otherwise, this is a brand-new application
         else:
             # convert the incoming application into the web form
-            form = MultiDict(xwalk.SuggestionFormXWalk.obj2form(ap))
+            form = MultiDict(suggestion_form.SuggestionFormXWalk.obj2form(ap))
 
             # create a template that will hold all the values we want to persist across the form submission
             template = models.Suggestion()
@@ -239,7 +239,7 @@ class ApplicationsCrudApi(CrudApi):
                 raise Api404Error()
 
             # convert the incoming application into the web form
-            form = MultiDict(xwalk.SuggestionFormXWalk.obj2form(new_ap))
+            form = MultiDict(suggestion_form.SuggestionFormXWalk.obj2form(new_ap))
 
             fc = formcontext.ApplicationFormFactory.get_form_context(role="publisher", form_data=form, source=vanilla_ap)
             if fc.validate():
@@ -265,7 +265,7 @@ class ApplicationsCrudApi(CrudApi):
                     raise Api404Error()
 
             # convert the incoming application into the web form
-            form = MultiDict(xwalk.SuggestionFormXWalk.obj2form(new_ap))
+            form = MultiDict(suggestion_form.SuggestionFormXWalk.obj2form(new_ap))
 
             fc = formcontext.ApplicationFormFactory.get_form_context(form_data=form, source=ap)
             if fc.validate():
@@ -324,7 +324,7 @@ class ApplicationsCrudApi(CrudApi):
         errors = fc.errors
         msg = "The following validation errors were received:\n"
         for fieldName, errorMessages in errors.items():
-            fieldName = xwalk.SuggestionFormXWalk.formField2objectField(fieldName)
+            fieldName = suggestion_form.SuggestionFormXWalk.formField2objectField(fieldName)
             msg += fieldName + " : " + "; ".join(errorMessages) + "\n"
         return msg
