+++ portality/formcontext/formcontext.py
@@ -5,12 +5,14 @@ from datetime import datetime
 from flask import render_template, url_for, request
 from flask_login import current_user
 
+import portality.formcontext.xwalks.admin_article_form
+import portality.formcontext.xwalks.journal_form
 from portality import constants
 from portality import models, app_email, util
 from portality.bll import DOAJ
-from portality.bll.exceptions import ArticleMergeConflict, DuplicateArticleException
 from portality.core import app
-from portality.formcontext import forms, xwalk, render, choices, emails, FormContextException
+from portality.formcontext import forms, render, choices, emails, FormContextException
+from portality.formcontext.xwalks import suggestion_form
 from portality.lcc import lcc_jstree
 from portality.ui.messages import Messages
 
@@ -554,10 +556,10 @@ class ApplicationContext(PrivateContext):
             if current_journal is not None:
                 cj = models.Journal.pull(current_journal)
                 if cj is not None:
-                    jform = xwalk.JournalFormXWalk.obj2form(cj)
+                    jform = portality.formcontext.xwalks.journal_form.JournalFormXWalk.obj2form(cj)
                     if "notes" in jform:
                         del jform["notes"]
-                    aform = xwalk.SuggestionFormXWalk.obj2form(self.source)
+                    aform = suggestion_form.SuggestionFormXWalk.obj2form(self.source)
                     if "notes" in aform:
                         del aform["notes"]
                     diff = self._form_diff(jform, aform)
@@ -643,7 +645,7 @@ class ManEdApplicationReview(ApplicationContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.ManEdApplicationReviewForm(data=xwalk.SuggestionFormXWalk.obj2form(self.source))
+        self.form = forms.ManEdApplicationReviewForm(data=suggestion_form.SuggestionFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
@@ -655,7 +657,7 @@ class ManEdApplicationReview(ApplicationContext):
         self._validate_editor_field()
 
     def form2target(self):
-        self.target = xwalk.SuggestionFormXWalk.form2obj(self.form)
+        self.target = suggestion_form.SuggestionFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -682,8 +684,8 @@ class ManEdApplicationReview(ApplicationContext):
 
         # FIXME: may want to factor this out of the suggestionformxwalk
         # If we have changed the editors assinged to this application, let them know.
-        is_editor_group_changed = xwalk.SuggestionFormXWalk.is_new_editor_group(self.form, self.source)
-        is_associate_editor_changed = xwalk.SuggestionFormXWalk.is_new_editor(self.form, self.source)
+        is_editor_group_changed = suggestion_form.SuggestionFormXWalk.is_new_editor_group(self.form, self.source)
+        is_associate_editor_changed = suggestion_form.SuggestionFormXWalk.is_new_editor(self.form, self.source)
 
         # record the event in the provenance tracker
         models.Provenance.make(current_user, "edit", self.target)
@@ -867,7 +869,7 @@ class EditorApplicationReview(ApplicationContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.EditorApplicationReviewForm(data=xwalk.SuggestionFormXWalk.obj2form(self.source))
+        self.form = forms.EditorApplicationReviewForm(data=suggestion_form.SuggestionFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
@@ -884,7 +886,7 @@ class EditorApplicationReview(ApplicationContext):
             self.form.application_status.data = constants.APPLICATION_STATUS_ACCEPTED
 
     def form2target(self):
-        self.target = xwalk.SuggestionFormXWalk.form2obj(self.form)
+        self.target = suggestion_form.SuggestionFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -911,7 +913,7 @@ class EditorApplicationReview(ApplicationContext):
         choices.Choices.validate_status_change('editor', self.source.application_status, self.target.application_status)
 
         # FIXME: may want to factor this out of the suggestionformxwalk
-        new_associate_assigned = xwalk.SuggestionFormXWalk.is_new_editor(self.form, self.source)
+        new_associate_assigned = suggestion_form.SuggestionFormXWalk.is_new_editor(self.form, self.source)
 
         # Save the target
         self.target.set_last_manual_update()
@@ -1026,7 +1028,7 @@ class AssEdApplicationReview(ApplicationContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.AssEdApplicationReviewForm(data=xwalk.SuggestionFormXWalk.obj2form(self.source))
+        self.form = forms.AssEdApplicationReviewForm(data=suggestion_form.SuggestionFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
@@ -1043,7 +1045,7 @@ class AssEdApplicationReview(ApplicationContext):
             self.form.application_status.data = constants.APPLICATION_STATUS_ACCEPTED
 
     def form2target(self):
-        self.target = xwalk.SuggestionFormXWalk.form2obj(self.form)
+        self.target = suggestion_form.SuggestionFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -1144,7 +1146,7 @@ class PublisherUpdateRequest(ApplicationContext):
         self._disable_fields()
 
     def source2form(self):
-        self.form = forms.PublisherUpdateRequestForm(data=xwalk.SuggestionFormXWalk.obj2form(self.source))
+        self.form = forms.PublisherUpdateRequestForm(data=suggestion_form.SuggestionFormXWalk.obj2form(self.source))
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
         self._disable_fields()
@@ -1183,7 +1185,7 @@ class PublisherUpdateRequest(ApplicationContext):
             self.form.confirm_contact_email.data = contact.get("email")
 
     def form2target(self):
-        self.target = xwalk.SuggestionFormXWalk.form2obj(self.form)
+        self.target = suggestion_form.SuggestionFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -1344,10 +1346,10 @@ class PublicApplication(ApplicationContext):
         self.form = forms.PublicApplicationForm(formdata=self.form_data)
 
     def source2form(self):
-        self.form = forms.PublicApplicationForm(data=xwalk.SuggestionFormXWalk.obj2form(self.source))
+        self.form = forms.PublicApplicationForm(data=suggestion_form.SuggestionFormXWalk.obj2form(self.source))
 
     def form2target(self):
-        self.target = xwalk.SuggestionFormXWalk.form2obj(self.form)
+        self.target = suggestion_form.SuggestionFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is not None:
@@ -1406,7 +1408,7 @@ class PublisherUpdateRequestReadOnly(PrivateContext):
         self.renderer.disable_all_fields(False)
 
     def source2form(self):
-        self.form = forms.PublisherUpdateRequestForm(data=xwalk.JournalFormXWalk.obj2form(self.source))
+        self.form = forms.PublisherUpdateRequestForm(data=portality.formcontext.xwalks.journal_form.JournalFormXWalk.obj2form(self.source))
         # self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
@@ -1471,7 +1473,7 @@ class ManEdJournalReview(PrivateContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.ManEdJournalReviewForm(data=xwalk.JournalFormXWalk.obj2form(self.source))
+        self.form = forms.ManEdJournalReviewForm(data=portality.formcontext.xwalks.journal_form.JournalFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
@@ -1481,7 +1483,7 @@ class ManEdJournalReview(PrivateContext):
         self._validate_editor_field()
 
     def form2target(self):
-        self.target = xwalk.JournalFormXWalk.form2obj(self.form)
+        self.target = portality.formcontext.xwalks.journal_form.JournalFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -1512,8 +1514,8 @@ class ManEdJournalReview(PrivateContext):
 
         # FIXME: may want to factor this out of the suggestionformxwalk
         # If we have changed the editors assinged to this application, let them know.
-        is_editor_group_changed = xwalk.JournalFormXWalk.is_new_editor_group(self.form, self.source)
-        is_associate_editor_changed = xwalk.JournalFormXWalk.is_new_editor(self.form, self.source)
+        is_editor_group_changed = portality.formcontext.xwalks.journal_form.JournalFormXWalk.is_new_editor_group(self.form, self.source)
+        is_associate_editor_changed = portality.formcontext.xwalks.journal_form.JournalFormXWalk.is_new_editor(self.form, self.source)
 
         # Save the target
         self.target.set_last_manual_update()
@@ -1596,13 +1598,13 @@ class EditorJournalReview(PrivateContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.EditorJournalReviewForm(data=xwalk.JournalFormXWalk.obj2form(self.source))
+        self.form = forms.EditorJournalReviewForm(data=portality.formcontext.xwalks.journal_form.JournalFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
 
     def form2target(self):
-        self.target = xwalk.JournalFormXWalk.form2obj(self.form)
+        self.target = portality.formcontext.xwalks.journal_form.JournalFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -1636,7 +1638,7 @@ class EditorJournalReview(PrivateContext):
         super(EditorJournalReview, self).finalise()
 
         # FIXME: may want to factor this out of the suggestionformxwalk
-        email_associate = xwalk.SuggestionFormXWalk.is_new_editor(self.form, self.source)
+        email_associate = suggestion_form.SuggestionFormXWalk.is_new_editor(self.form, self.source)
 
         # Save the target
         self.target.set_last_manual_update()
@@ -1674,13 +1676,13 @@ class AssEdJournalReview(PrivateContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.AssEdJournalReviewForm(data=xwalk.JournalFormXWalk.obj2form(self.source))
+        self.form = forms.AssEdJournalReviewForm(data=portality.formcontext.xwalks.journal_form.JournalFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
 
     def form2target(self):
-        self.target = xwalk.JournalFormXWalk.form2obj(self.form)
+        self.target = portality.formcontext.xwalks.journal_form.JournalFormXWalk.form2obj(self.form)
 
     def patch_target(self):
         if self.source is None:
@@ -1743,7 +1745,7 @@ class ReadOnlyJournal(PrivateContext):
         self._expand_url_descriptions(URL_FIELDS)
 
     def source2form(self):
-        self.form = forms.ReadOnlyJournalForm(data=xwalk.JournalFormXWalk.obj2form(self.source))
+        self.form = forms.ReadOnlyJournalForm(data=portality.formcontext.xwalks.journal_form.JournalFormXWalk.obj2form(self.source))
         self._set_choices()
         self._expand_descriptions(FIELDS_WITH_DESCRIPTION)
         self._expand_url_descriptions(URL_FIELDS)
@@ -1812,7 +1814,7 @@ class AdminArticleForm(FormContext):
 
     def source2form(self):
         self.form = forms.ArticleForm()
-        xwalk.AdminArticleXwalk.obj2form(self.form, article=self.source)
+        portality.formcontext.xwalks.admin_article_form.AdminArticleFormXwalk.obj2form(self.form, article=self.source)
         self._set_choices()
 
     def data2form(self):
@@ -1820,7 +1822,7 @@ class AdminArticleForm(FormContext):
         self._set_choices()
 
     def form2target(self):
-        self.target = xwalk.AdminArticleXwalk.form2obj(form=self.form)
+        self.target = portality.formcontext.xwalks.admin_article_form.AdminArticleFormXwalk.form2obj(form=self.form)
 
     def render_template(self, **kwargs):
         if "more_authors" in kwargs and kwargs["more_authors"] == True:
