+++ portality/view/admin.py
@@ -162,7 +162,7 @@ def article_page(article_id):
         return fc.render_template()
 
     elif request.method == "POST":
-        user = models.Account.pull(current_user.id)
+        user = current_user._get_current_object()
         fc = formcontext.ArticleFormFactory.get_from_context(role="admin", source=ap, user=user, form_data=request.form)
         # first we need to do any server-side form modifications which
         # the user might request by pressing the add/remove authors buttons
